docs: add more jsdocs and optimize functions for manifest
Add more jsdocs to prep the API doc generation.

Add event display to API docs.

Discovered in looking over the manifest that it can support
knowing exported functions if they are declared functions
and not variable assigned ones. So tweaked the ESLint config
to allow for that operation to happen.