refactor(tree): do not use virtualizer element
Utilize the directive while also removing the Shadow DOM.
This reflects what is done in the virtualizer itself.
As the Shadow DOM is not necessary to encapsulate any
styling since none exists here that external developers
shouldn't have access to already. This minimizes the total
DOM impact and should provide consumers a better experience.