docs(a11y): initial content
Provide documentation for accessibility guidance that has been
implemented in the package.

Some tweaks to API exposure were made to adjust to match language
used in the accessibility document for consistent terminology.

Tweaked the aria roles to use internals as well so the baseline
is set by the component and developers may optionally override
if they chose in user-land.